refactor: split style color and dimensions
